refactor: move checkPairsShuffled back into a separate function
