+++ app/src/espresso/java/xyz/lbres/trickcalculator/testutils/helpers.kt
@@ -1,5 +1,12 @@
 package xyz.lbres.trickcalculator.testutils
 
+/**
+ * Alternate to standard [repeat] function, to allow breaking out of the loop.
+ *
+ * @param maxRepeats [Int]: number of times to repeat code without breaking
+ * @param breakCondition () -> [Boolean]: condition to break out of loop
+ * @param code () -> [Unit]: code to execute repeatedly
+ */
 fun repeatUntil(maxRepeats: Int, breakCondition: () -> Boolean, code: () -> Unit) {
     var repeats = 0
     while (repeats < maxRepeats && !breakCondition()) {
