+++ app/src/espresso/java/xyz/lbres/trickcalculator/ui/history/historyHelpers.kt
@@ -3,19 +3,14 @@ package xyz.lbres.trickcalculator.ui.history
 import android.view.View
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.withChild
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.allOf
-import org.hamcrest.Matchers.anyOf
-import org.hamcrest.Matchers.not
 import xyz.lbres.trickcalculator.R
 import xyz.lbres.trickcalculator.testutils.closeFragment
-import xyz.lbres.trickcalculator.testutils.matchers.withViewHolder
 import xyz.lbres.trickcalculator.testutils.openSettingsFragment
-import xyz.lbres.trickcalculator.testutils.viewactions.scrollToPosition
 
 /**
  * Representation of a history item displayed in the UI.
@@ -33,34 +28,6 @@ val withHistoryItem: (String, String) -> Matcher<View?> = { computation: String,
         withChild(withText(computation)),
         withChild(withChild(withText(result)))
     )
-}
-
-/**
- * Check that the values in a view holder match one of the items in the compute history
- *
- * @param position [Int]: position of view holder to check
- * @param computeHistory [TestHistory]: list of items in history
- */
-fun checkViewHolderInHistory(position: Int, computeHistory: TestHistory) {
-    val recyclerId = R.id.itemsRecycler
-    val historyMatcher = anyOf(computeHistory.map { withHistoryItem(it.first, it.second) })
-
-    onView(withId(recyclerId)).perform(scrollToPosition(position))
-    onView(withViewHolder(recyclerId, position)).check(matches(historyMatcher))
-}
-
-/**
- * Check that the values in a view holder do not match any of the items in the compute history
- *
- * @param position [Int]: position of view holder to check
- * @param computeHistory [TestHistory]: list of items in history
- */
-fun checkViewHolderNotInHistory(position: Int, computeHistory: TestHistory) {
-    val recyclerId = R.id.itemsRecycler
-    val historyMatcher = anyOf(computeHistory.map { withHistoryItem(it.first, it.second) })
-
-    onView(withId(recyclerId)).perform(scrollToPosition(position))
-    onView(withViewHolder(recyclerId, position)).check(matches(not(historyMatcher)))
 }
 
 /**
