+++ app/src/espresso/java/xyz/lbres/trickcalculator/ui/history/testRandomness2.kt
@@ -1,21 +1,22 @@
 package xyz.lbres.trickcalculator.ui.history
 
-import android.view.View
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withChild
+import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
-import org.hamcrest.Matcher
 import org.hamcrest.Matchers.allOf
 import org.hamcrest.Matchers.anyOf
 import org.hamcrest.Matchers.not
 import xyz.lbres.kotlinutils.general.ternaryIf
 import xyz.lbres.trickcalculator.R
 import xyz.lbres.trickcalculator.testutils.closeFragment
+import xyz.lbres.trickcalculator.testutils.matchers.withViewHolder
 import xyz.lbres.trickcalculator.testutils.openHistoryFragment
 import xyz.lbres.trickcalculator.testutils.repeatUntil
 import xyz.lbres.trickcalculator.testutils.toggleShuffleOperators
+import xyz.lbres.trickcalculator.testutils.viewactions.scrollToPosition
 import xyz.lbres.trickcalculator.ui.calculator.clearText
 import xyz.lbres.trickcalculator.ui.calculator.equals
 import xyz.lbres.trickcalculator.ui.calculator.typeText
@@ -84,13 +85,14 @@ fun testRandomness2() {
  * @param computeHistory [TestHistory]: list of items in history
  */
 private fun checkCorrectData(computeHistory: TestHistory) {
-    var shuffled = false
+    var orderShuffled = false
+    var pairsShuffled = false
     val historySize = computeHistory.size
 
     // additional repeats for 2 items due to occasional failures
     val repeatCount = ternaryIf(historySize == 2, 10, 5)
 
-    repeatUntil(repeatCount, { shuffled }) {
+    repeatUntil(repeatCount, { orderShuffled && pairsShuffled }) {
         openHistoryFragment()
 
         // check that all items are displayed
@@ -106,25 +108,46 @@ private fun checkCorrectData(computeHistory: TestHistory) {
             { it.map { it.second } },
             { result -> withChild(withChild(withText(result as String))) }
         )
-        // matching computation/result strings not displayed together
-        val shuffledPairsCheck = ShuffledCheckInfo({ it }, { _ ->
-            val computeTextMatcher = anyOf(computeHistory.map { withChild(withText(it.first)) })
-            val resultTextMatcher = anyOf(computeHistory.map { withChild(withChild(withText(it.second))) })
-            var historyMatcher: Matcher<View?> = anyOf(computeHistory.map { withHistoryItem(it.first, it.second) })
-            // regular matcher creates issues when size = 2
-            if (computeHistory.size == 2) {
-                historyMatcher = withHistoryItem(computeHistory[0].first, computeHistory[1].second)
-            }
 
-            allOf(computeTextMatcher, resultTextMatcher, not(historyMatcher))
-        })
-
-        shuffled = shuffled || checkItemsShuffled(computeHistory, listOf(shuffledComputationCheck, shuffledResultCheck, shuffledPairsCheck))
+        orderShuffled = orderShuffled || checkItemsShuffled(computeHistory, listOf(shuffledComputationCheck, shuffledResultCheck))
+        pairsShuffled = pairsShuffled || checkPairsShuffled(computeHistory)
 
         closeFragment()
     }
 
-    if (historySize > 1 && !shuffled) {
+    if (historySize > 1 && !(orderShuffled && pairsShuffled)) {
         throw AssertionError("History items and pairs should be shuffled in history randomness 2. History: $computeHistory")
     }
+}
+
+/**
+ * Determine if the computation/result matches have been shuffled
+ *
+ * @param computeHistory [TestHistory]: list of items in history
+ * @return [Boolean]: true if at least one viewholder contains a mismatched computation/result, false if all pairs match
+ */
+private fun checkPairsShuffled(computeHistory: TestHistory): Boolean {
+    val recyclerId = R.id.itemsRecycler
+    val historySize = computeHistory.size
+    if (historySize < 2) {
+        return true
+    }
+
+    val computeTextMatcher = anyOf(computeHistory.map { withChild(withText(it.first)) })
+    val resultTextMatcher = anyOf(computeHistory.map { withChild(withChild(withText(it.second))) })
+    val historyMatcher = anyOf(computeHistory.map { withHistoryItem(it.first, it.second) })
+
+    for (position in 0 until historySize) {
+        try {
+            onView(withId(recyclerId)).perform(scrollToPosition(position))
+            // both values are part of the history, but value doesn't match a real item
+            onView(withViewHolder(recyclerId, position)).check(
+                matches(allOf(computeTextMatcher, resultTextMatcher, not(historyMatcher)))
+            )
+
+            return true
+        } catch (_: Throwable) {}
+    }
+
+    return false
 }
